www/api/requirements.txt
Django==3.0
psycopg2==2.8.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.4.0
django-filter==2.2.0
python-decouple==3.3
requests==2.23.0
cryptography~=2.8
drf-jwt~=1.13.3
pyjwt~=1.7.1
django-cors-headers==3.4.0